requirements.txt
beautifulsoup4==4.3.2
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11
django-braces==1.11.0
django-cors-headers==2.1.0
django-crispy-forms==1.4.0
django-durationfield==0.5.1
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.0.8
django-toolkit==0.3.3
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.6
lxml==3.4.4
oauthlib==2.0.2
psycopg2==2.7.3
PyJWT==1.5.2
python-dateutil==2.4.2
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
tldextract==1.6
unicodecsv==0.13.0
urllib3==1.22
whitenoise==3.3.0
